+++ discord/ui/action_row.py
@@ -67,6 +67,7 @@ if TYPE_CHECKING:
     from ..components import SelectOption
     from ..interactions import Interaction
     from .container import Container
+    from .dynamic import DynamicItem
 
     SelectCallbackDecorator = Callable[[ItemCallbackType['S', BaseSelectT]], BaseSelectT]
 
@@ -194,9 +195,18 @@ class ActionRow(Item[V]):
         # it should error anyways.
         return True
 
+    def _swap_item(self, base: Item, new: DynamicItem, custom_id: str) -> None:
+        child_index = self._children.index(base)
+        self._children[child_index] = new  # type: ignore
+
     @property
     def width(self):
         return 5
+
+    @property
+    def _total_count(self) -> int:
+        # 1 for self and all children
+        return 1 + len(self._children)
 
     @property
     def type(self) -> Literal[ComponentType.action_row]:
@@ -348,6 +358,7 @@ class ActionRow(Item[V]):
         The function being decorated should have three parameters, ``self`` representing
         the :class:`discord.ui.ActionRow`, the :class:`discord.Interaction` you receive and
         the :class:`discord.ui.Button` being pressed.
+
         .. note::
 
             Buttons with a URL or a SKU cannot be created with this function.